Conversion to run on SIM1 computer
Converted program to run on the SIM1 linux computer and use the new
symlinks to the data files.  Modified the search routine to use the
linux find command and had to tweak a couple of other things to get it
to work.  Still have some tweaks to the UI windows to do.